FE: Rename side-nav/Boards -> WorkspaceBoards
