+++ frontend/src/lib/figma/navigation/side-nav/WorkspaceBoards.svelte
