+++ frontend/src/stories/figma/navigation/side-nav/WorkspaceBoards.stories.ts
@@ -1,6 +1,6 @@
 // SPDX-License-Identifier: AGPL-3.0-or-later
 /*
- *  Copyright (C) 2023 JWP Consulting GK
+ *  Copyright (C) 2023-2024 JWP Consulting GK
  *
  *  This program is free software: you can redistribute it and/or modify
  *  it under the terms of the GNU Affero General Public License as published
@@ -17,7 +17,7 @@
  */
 import type { Meta, StoryObj } from "@storybook/svelte";
 
-import Boards from "$lib/figma/navigation/side-nav/Boards.svelte";
+import Boards from "$lib/figma/navigation/side-nav/WorkspaceBoards.svelte";
 import { workspace } from "$lib/storybook";
 
 const meta: Meta<Boards> = {
